refactor: change class package
