+++ adcio_agent/src/test/java/ai/corca/adcio_agent/AgentUnitTest.kt
@@ -1,5 +1,6 @@
 package ai.corca.adcio_agent
 
+import ai.corca.adcio_agent.provider.WebViewManager
 import org.junit.Assert.assertEquals
 import org.junit.Test
